fix: rename text area
