+++ src/components/candidate/Popups/PopupDescription/PopupDescription.jsx
@@ -1,5 +1,5 @@
 import { ButtonPrimary } from '@/components/shared/ButtonPrimary';
-import { TextArea } from '@/components/shared/TextArea/TextArea';
+import { TextArea } from '@/components/shared/TextArea';
 import { withTheme } from '@/contexts/ThemeContext';
 
 const PopupDescription = withTheme(({ className, variant = 'default', ...props }) => {
