+++ src/components/candidate/Popups/PopupEducation/PopupEducation.jsx
@@ -1,7 +1,7 @@
 import { ButtonPrimary } from '@/components/shared/ButtonPrimary';
 import { DataPicker } from '@/components/shared/DataPicker';
 import { InputLabel } from '@/components/shared/InputLabel';
-import { TextArea } from '@/components/shared/TextArea/TextArea';
+import { TextArea } from '@/components/shared/TextArea/';
 import { withTheme } from '@/contexts/ThemeContext';
 
 const PopupEducation = withTheme(({ className, variant = 'default', editItem, ...props }) => {
