+++ src/components/candidate/Popups/PopupExperiences/PopupExperiences.jsx
@@ -1,7 +1,7 @@
 import { ButtonPrimary } from '@/components/shared/ButtonPrimary';
 import { DataPicker } from '@/components/shared/DataPicker/DataPicker';
 import { InputLabel } from '@/components/shared/InputLabel';
-import { TextArea } from '@/components/shared/TextArea/TextArea';
+import { TextArea } from '@/components/shared/TextArea';
 import { withTheme } from '@/contexts/ThemeContext';
 
 const PopupExperiences = withTheme(({ className, variant = 'default', editItem, ...props }) => {
